Post-HTML-Validator-Check-2
I thought I had originally ‘completed’ Assignment 2, but this was
before undertaking lessons four and five. Therefore, I have gone back
and tweaked a few things based on what was taught in these lessons, as
well as utilising a couple of the handy tips/links that were made
available to us in Schoology.

At this stage, I don’t think I have covered everything, but I’m
planning to revisit it when I make a start on Assignment 3.